refactor: style 폴더 분리
