+++ src/components/ImgWrapper/ImgWrapper.jsx
@@ -1,29 +1,17 @@
-import styled from 'styled-components';
+import Styled from './style';
 import PropTypes from 'prop-types';
 
-const ImgWrapper = ({ src }) => {
+const ImgWrapper = ({ src, alt = '이미지' }) => {
   return (
     <Styled.Wrapper>
-      <Styled.Img src={src} alt="에러이미지" />
+      <Styled.Img src={src} alt={alt} />
     </Styled.Wrapper>
   );
 };
 
 ImgWrapper.propTypes = {
   src: PropTypes.string,
-};
-
-const Styled = {
-  Wrapper: styled.div`
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  `,
-  Img: styled.img`
-    width: 60%;
-  `,
+  alt: PropTypes.string,
 };
 
 export default ImgWrapper;
