+++ src/components/ImgWrapper/style.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+const Styled = {
+  Wrapper: styled.div`
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  `,
+  Img: styled.img`
+    width: 60%;
+  `,
+};
+
+export default Styled;
