+++ src/components/MenuItem/style.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+const Styled = {
+  Wrapper: styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-weight: 500;
+    font-size: 24px;
+    color: #ffffff;
+    cursor: pointer;
+  `,
+};
+
+export default Styled;
