+++ src/components/ProductDetail/ProductDetail.jsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import Styled from './style';
 import PropTypes from 'prop-types';
 import Button from 'components/Button/Button';
 
@@ -27,32 +27,6 @@ ProductDetail.propTypes = {
   name: PropTypes.string,
   price: PropTypes.string,
   onClick: PropTypes.func,
-};
-
-const Styled = {
-  ProductImg: styled.img`
-    width: 350px;
-    height: 350px;
-  `,
-  ProductInfo: styled.div`
-    width: 380px;
-  `,
-  ProductName: styled.div`
-    margin: 10px 0 10px 15px;
-    font-size: 18px;
-    font-weight: 700;
-  `,
-  ProductPrice: styled.div`
-    margin: 0 0 10px 15px;
-    display: flex;
-    justify-content: space-between;
-    font-size: 16px;
-    font-weight: 400;
-  `,
-
-  Line: styled.hr`
-    margin: 0;
-  `,
 };
 
 export default ProductDetail;
