+++ src/pages/Cart/style.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const Styled = {
+  Wrapper: styled.section`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    margin: 50px;
+  `,
+};
+
+export default Styled;
